+++ src/main/java/wooteco/PageController.java
@@ -1,4 +1,4 @@
-package wooteco.subway;
+package wooteco;
 
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
